{{cookiecutter.repo_name}}/requirements/base.txt
{%- if cookiecutter.use_celery == "y" -%}
celery==3.1.23
{%- endif %}
Django==1.10.1
{%- if cookiecutter.use_allauth == 'y' %}
django-allauth==0.27.0
{%- endif %}
django-braces==1.9.0
django-environ==0.4.0
django-extensions==1.7.4
django-model-utils==2.5.2
Pillow==3.3.1
psycopg2==2.6.2
pytz==2016.6.1
raven==5.26.0
{%- if cookiecutter.use_redis == "y" %}
redis>=2.10.0
django-redis-cache==1.7.1
{%- endif %}
{%- if cookiecutter.use_robots == "y" %}
django-robots==2.0
{%- endif %}
django-solo==1.1.2
